+++ src/solutions/day5.rs
@@ -7,14 +7,14 @@ pub fn main() {
     part_2(prog);
 }
 
-fn part_1(prog: Vec<i32>) {
+fn part_1(prog: Vec<i64>) {
     let mut cpu = IntcodeComputer::new(prog);
     cpu.run_io(Box::new(|| 1), Box::new(|x| if x != 0 {
             println!("{}", x)
     }));
 }
 
-fn part_2(prog: Vec<i32>) {
+fn part_2(prog: Vec<i64>) {
     let mut cpu = IntcodeComputer::new(prog);
     cpu.run_io(Box::new(|| 5), Box::new(|x| println!("{}", x)));
 }
